Bugs fixed

+++ snippet_app/views.py
@@ -6,6 +6,7 @@ from rest_framework.response import Response
 from .serializers import TagSerializer, UserSerializer, TextSerializer
 from .models import Text, Tag
 from django.contrib.auth.hashers import make_password, check_password
+from functools import wraps
 import datetime, jwt, os
 from binascii import hexlify
 
@@ -58,50 +59,62 @@ class Snippet(APIView):
     def get(self, request):
         token = request.COOKIES.get('jwt')
         if not token:
-            raise AuthenticationFailed('UnAuthenticate!')
+            raise AuthenticationFailed('UnAuthenticated!')
         try:
-            payload = jwt.decode(token, secret_key, algorithms=['HS256'])
+            snippets = Text.objects.all()
+            serializer = TextSerializer(snippets, many=True)
+            return Response(serializer.data)
         except jwt.ExpiredSignatureError:
             raise AuthenticationFailed('UnAuthenticated!')
-        snippets = Text.objects.all()
-        serializer = TextSerializer(snippets, many=True)
-        return Response(serializer.data)
+        
     
     def post(self, request):
-        text = request.data['text']
-        title = request.data['title']
-        if Text.objects.filter(title=title).exists():
-            return Response({
-                "message" : "The title already exists"
-            })
-        if User.objects.filter(id=request.data['user']).exists():
-            user = User.objects.get(id=request.data['user'])
-        else:
-            return Response({
-                "message" : "The user doesn't exist"
-            })
-        if 'tag' in request.data.keys():
-            tag = request.data['tag']
-            if Tag.objects.filter(tag=tag).exists():
+        token = request.COOKIES.get('jwt')
+        if not token:
+            raise AuthenticationFailed('UnAuthenticated!')
+        try:
+            text = request.data['text']
+            title = request.data['title']
+            if Text.objects.filter(title=title).exists():
                 return Response({
-                    "message" : "The tag already exists"
+                    "message" : "The title already exists"
                 })
-            curr_tag = Tag.objects.filter(tag=tag).first()
-            Text.objects.create(text=text, title=title, tag=curr_tag, user=user)
-            curr_text = Text.objects.filter(title=title).first()
-            Tag.objects.create(tag=tag, text=curr_text)
+            if User.objects.filter(id=request.data['user']).exists():
+                user = User.objects.get(id=request.data['user'])
+            else:
+                return Response({
+                    "message" : "The user doesn't exist"
+                })
+            if 'tag' in request.data.keys():
+                tag = request.data['tag']
+                if Tag.objects.filter(tag=tag).exists():
+                    return Response({
+                        "message" : "The tag already exists"
+                    })
+                curr_tag = Tag.objects.filter(tag=tag).first()
+                Text.objects.create(text=text, title=title, tag=curr_tag, user=user)
+                curr_text = Text.objects.filter(title=title).first()
+                Tag.objects.create(tag=tag, text=curr_text)
+                return Response(status=status.HTTP_201_CREATED)
+            Text.objects.create(text=text, title=title, user=user)
             return Response(status=status.HTTP_201_CREATED)
-        Text.objects.create(text=text, title=title, user=user)
-        return Response(status=status.HTTP_201_CREATED)
+        except jwt.ExpiredSignatureError:
+            raise AuthenticationFailed('UnAuthenticated!')
+        
     
-
 #this class is for to get the count of the snippets thar are stored on the database.
 class CountSnippets(APIView):
     def get(self, request):
-        count = Text.objects.all().count()
-        return Response({
-            "count of snippets" : count
-            })
+        token = request.COOKIES.get('jwt')
+        if not token:
+            raise AuthenticationFailed('UnAuthenticated!')
+        try:
+            count = Text.objects.all().count()
+            return Response({
+                "count of snippets" : count
+                })
+        except jwt.ExpiredSignatureError:
+            raise AuthenticationFailed('UnAuthenticated!')
        
         
 """" The class for to access the specific data snippet.
@@ -111,37 +124,71 @@ The delete function is used to delete the particular snippet."""
 class SnippetSpecific(APIView):
     
     def get(self, request, title):
-        snippet = Text.objects.get(title=title)
-        serializer = TextSerializer(snippet, many=False)
-        return Response(serializer.data)
+        token = request.COOKIES.get('jwt')
+        if not token:
+            raise AuthenticationFailed('UnAuthenticated!')
+        try:
+            snippet = Text.objects.filter(title=title).first()
+            if snippet is not None:
+                serializer = TextSerializer(snippet, many=False)
+                return Response(serializer.data)
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        except jwt.ExpiredSignatureError:
+            raise AuthenticationFailed('UnAuthenticated!')
     
     def put(self, request, title):
-        snippet = Text.objects.get(title=title)
-        serializer = TextSerializer(snippet, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data)
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        token = request.COOKIES.get('jwt')
+        if not token:
+            raise AuthenticationFailed('UnAuthenticated!')
+        try:
+            snippet = Text.objects.filter(title=title).first()
+            if snippet is not None:
+                serializer = TextSerializer(snippet, data=request.data)
+                if serializer.is_valid():
+                    serializer.save()
+                    return Response(serializer.data)
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response(status=status.HTTP_404_NOT_FOUND)
+        except jwt.ExpiredSignatureError:
+            raise AuthenticationFailed('UnAuthenticated!')
     
     def delete(self, request, title):
-        snippet = Text.objects.get(title=title)
-        snippet.delete()
-        return Response({
-            "message" : "The snippet has deleted successfully"
-        })
+        token = request.COOKIES.get('jwt')
+        if not token:
+            raise AuthenticationFailed('UnAuthenticated!')
+        try:
+            snippet = Text.objects.get(title=title)
+            snippet.delete()
+            return Response({
+                "message" : "The snippet has deleted successfully"
+            })
+        except jwt.ExpiredSignatureError:
+            raise AuthenticationFailed('UnAuthenticated!')
         
 
 #this class is for to list the tags are stored on the database.
 class ListTags(APIView):
     def get(self, request):
-        tags = Tag.objects.all()
-        serializer = TagSerializer(tags, many=True)
-        return Response(serializer.data)
+        token = request.COOKIES.get('jwt')
+        if not token:
+            raise AuthenticationFailed('UnAuthenticated!')
+        try:
+            tags = Tag.objects.all()
+            serializer = TagSerializer(tags, many=True)
+            return Response(serializer.data)            
+        except jwt.ExpiredSignatureError:
+            raise AuthenticationFailed('UnAuthenticated!')
     
 
 #this class is for to get the particular tag.
 class SpecificTag(APIView):
     def get(self, request, tag):
-        tag = Tag.objects.get(tag=tag)
-        serializer = TagSerializer(tag, many=False)
-        return Response(serializer.data)
+        token = request.COOKIES.get('jwt')
+        if not token:
+            raise AuthenticationFailed('UnAuthenticated!')
+        try:
+            tag = Tag.objects.get(tag=tag)
+            serializer = TagSerializer(tag, many=False)
+            return Response(serializer.data)            
+        except jwt.ExpiredSignatureError:
+            raise AuthenticationFailed('UnAuthenticated!')
